Move IconButton to components folder;
